vector-database/requirements.txt
flask
numpy
faiss-cpu
